updated usePosition and useWidth; first and last bars in system are now correctly calculated
