+++ components/Barline/Barline.js
@@ -7,32 +7,60 @@ import metadata from './../../public/fonts/bravura/bravura_metadata.json';
 import styles from './Barline.module.css';
 import { withNoSSR } from '../../hooks/withNoSSR';
 
-function Barline({ children, final = false, column, row = '1 / -1', separation }) {
-  const barlineRef = useRef(null);
+function RegularBarline() {
+  return <rect x={0} y={0} width={1} height={1} />;
+}
+
+function FinalBarline() {
+  return (
+    <>
+      <rect
+        x={0}
+        y={0}
+        width={metadata.engravingDefaults.thinBarlineThickness / 4}
+        height={1}
+      />
+      <rect
+        x={
+          (
+            metadata.engravingDefaults.thinBarlineThickness
+            + metadata.engravingDefaults.barlineSeparation
+          ) / 4
+        }
+        y={0}
+        width={metadata.engravingDefaults.thickBarlineThickness / 4}
+        height={1}
+      />
+    </>
+  )
+}
+
+function Barline({ children, type = 'regular', column, row = '1 / -1', separation }) {
+  const ref = useRef(null);
   const className = useVars({
-    varRef: barlineRef,
+    varRef: ref,
     defaultStyles: [styles.barline],
     conditionalStyles: [{ condition: separation, operator: "&&", style: styles.separation }],
     key: '--separation',
     value: `${metadata.engravingDefaults.barlineSeparation / 4}rem`,
   })
   useVars({
-    varRef: barlineRef,
+    varRef: ref,
     key: '--column',
-    value: final ? 'me-bar / m-end' : column,
+    value: type === "final" ? 'me-bar / m-end' : column,
   })
   useVars({
-    varRef: barlineRef,
+    varRef: ref,
     key: '--row',
     value: row,
   })
 
   return (
     <svg
-      ref={barlineRef}
-      viewBox={final ? `0 0 ${(metadata.engravingDefaults.thinBarlineThickness + metadata.engravingDefaults.barlineSeparation + metadata.engravingDefaults.thickBarlineThickness) / 4} 1`
+      ref={ref}
+      viewBox={type === "final" ? `0 0 ${(metadata.engravingDefaults.thinBarlineThickness + metadata.engravingDefaults.barlineSeparation + metadata.engravingDefaults.thickBarlineThickness) / 4} 1`
         : `0 0 1 1`}
-      width={final ? `${(metadata.engravingDefaults.thinBarlineThickness + metadata.engravingDefaults.barlineSeparation + metadata.engravingDefaults.thickBarlineThickness) / 4}rem`
+      width={type === "final" ? `${(metadata.engravingDefaults.thinBarlineThickness + metadata.engravingDefaults.barlineSeparation + metadata.engravingDefaults.thickBarlineThickness) / 4}rem`
         : `${metadata.engravingDefaults.thinBarlineThickness / 4}rem`}
       height="100%"
       preserveAspectRatio='none'
@@ -40,29 +68,7 @@ function Barline({ children, final = false, column, row = '1 / -1', separation }
       className={clsx(className)}
     >
       {children}
-      {final ? (
-        <>
-          <rect
-            x={0}
-            y={0}
-            width={metadata.engravingDefaults.thinBarlineThickness / 4}
-            height={1}
-          />
-          <rect
-            x={
-              (
-                metadata.engravingDefaults.thinBarlineThickness
-                + metadata.engravingDefaults.barlineSeparation
-              ) / 4
-            }
-            y={0}
-            width={metadata.engravingDefaults.thickBarlineThickness / 4}
-            height={1}
-          />
-        </>
-      ) : (
-        <rect x={0} y={0} width={1} height={1} />
-       )}
+      {type === "final" ? <FinalBarline /> : <RegularBarline />}
     </svg>
   );
 }
