+++ components/Staff/Staff.module.css
@@ -1,0 +1,31 @@
+.staff {
+  --number: 1;
+  --pitches: [b4] 0.125rem;
+  --start: m-start;
+  --end: m-end;
+  --transformY: 0rem;
+  grid-row: staff var(--number) / span 1;
+  display: grid;
+  grid-template-columns: subgrid;
+  grid-template-rows: var(--pitches);
+  align-content: start;
+  height: 1rem;
+  grid-column: var(--start) / var(--end);
+  transform: translateY(var(--transformY));
+}
+.staff .line {
+  --pitch: b4;
+  --start: m-start;
+  --end: m-end;
+  --weight: 0.0325rem;
+  align-self: start;
+  width: 100%;
+  height: 0;
+  border-style: solid;
+  border-top: var(--weight) solid currentColor;
+  border-bottom: none;
+  margin: 0;
+  transform: translateY(-50%);
+  grid-row: var(--pitch) / var(--pitch);
+  grid-column: var(--start) / var(--end);
+}
