+++ components/Systems/Systems.module.css
@@ -9,4 +9,5 @@
   height: 100%;
   min-height: fit-content;
   margin: 2rem;
+  background: white;
 }
