+++ fixtures/2024-08-29--test.json
@@ -17,6 +17,14 @@
   },
   "parts": [
     {
+      "clefs": [
+        {
+          "clef": {
+            "sign": "G",
+            "staffPosition": -2
+          }
+        }
+      ],
       "measures": [
         {
           "clefs": [
